+++ mustache_lambda.cpp
@@ -8,6 +8,7 @@
 #include "php_mustache.h"
 #include "mustache_private.hpp"
 #include <algorithm>
+#include "mustache_exceptions.hpp"
 #include "mustache_lambda_helper.hpp"
 #include "mustache_lambda.hpp"
 
@@ -27,7 +28,8 @@ std::string Lambda::invoke()
   return closure_result_string;
 }
 
-std::string Lambda::invoke(std::string * text, mustache::Renderer * renderer)
+std::string Lambda::invoke(
+    std::string_view text, mustache::LambdaRenderContext context)
 {
   zval closure_result;
   std::string closure_result_string;
@@ -40,13 +42,16 @@ std::string Lambda::invoke(std::string * text, mustache::Renderer * renderer)
     params = (zval *) safe_emalloc(sizeof(zval), param_count, 0);
   }
   if( param_count >= 1 ) {
-    ZVAL_STRINGL(&params[0], text->c_str(), text->length());
+    ZVAL_STRINGL(&params[0], text.data(), text.size());
   }
   if( param_count >= 2 ) {
     object_init_ex(&params[1], MustacheLambdaHelper_ce_ptr);
 
     struct php_obj_MustacheLambdaHelper * payload = php_mustache_lambda_helper_object_fetch_object(&params[1]);
-    payload->renderer = renderer;
+    if( payload->state == NULL ) {
+      throw InvalidParameterException("MustacheLambdaHelper state was not initialized properly");
+    }
+    payload->state->context = context;
   }
 
   if( invokeUserFunction(&closure_result, param_count, params) == SUCCESS ) {