+++ mustache_lambda.hpp
@@ -3,15 +3,16 @@
 #define PHP_MUSTACHE_LAMBDA_HPP
 
 #include <mustache/lambda.hpp>
+#include <string_view>
 
 class Lambda : public mustache::Lambda {
   protected:
     virtual int getUserFunctionParamCount() = 0;
     virtual int invokeUserFunction(zval *retval_ptr, int param_count, zval params[]) = 0;
   public:
-    std::string invoke();
-    std::string invoke(std::string * text, mustache::Renderer * renderer);
+    std::string invoke() override;
+    std::string invoke(
+        std::string_view text, mustache::LambdaRenderContext context) override;
 };
 
 #endif /* PHP_MUSTACHE_LAMBDA_HPP */
-